refactor: Modifica a tipagem
Cria um arquivo types que contém todas as tipagens usadas no projeto.
