+++ src/store/index.ts
@@ -1,10 +1,5 @@
 import { createStore, combineReducers } from 'redux';
 import { GlobalProducts } from './modules/products/reducers';
-import { AllProducts } from './modules/products/types';
-
-export interface IRootStore {
-    GlobalProducts: AllProducts
-};
 
 const combineStore = combineReducers({
     GlobalProducts
