+++ src/store/modules/products/types/index.ts
@@ -1,20 +1,2 @@
 export const ADD_PRODUCT = 'ADD_PRODUCT';
 export const REMOVE_PRODUCT = 'REMOVE_PRODUCT';
-
-export interface Product {  
-    id: string,
-    title: string,
-    price: string,
-    image: string,
-    quantity: string
-}
-
-export interface CartProduct extends Product {
-    countCart: number
-}
-
-export interface AllProducts {
-    products: CartProduct[],
-    totalItemCart: number,
-    totalPriceCart: number
-}
